requirements.txt
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
click==7.1.2
colorama==0.4.3
dash
dash_bootstrap_components
dash_core_components
dash_html_components
Flask==1.1.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
numpy==1.19
matplotlib==3.3.3
mccabe==0.6.1
more-itertools==8.2.0
openpyxl
packaging==20.3
pandas==1.2.1
pluggy==0.13.1
plotly_express
py==1.8.1
pylint==2.4.4
pyparsing==2.4.7
pytest==5.4.1
six==1.14.0
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.11.2
xlrd
